Ticket 3435 - Moved out PropertyValueMergeService-specifc unit tests to a separate class.
